requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg[binary]==3.2.2
alembic==1.14.0
pydantic==1.10.18
qrcode==7.4.2
pillow==10.4.0
python-multipart==0.0.9
apscheduler==3.10.4
sentry-sdk==2.13.0
pytest==8.3.3
httpx==0.27.2
pytest-cov==5.0.0